Replaced duration option with speed option
Depreciated duration option in favor of speed. This gives more control
as the plugin will calculate how many characters in the string and
determine a duration based on that and the speed you have given it. This
allows it to be more flexible in that the same code can be used for long
or short strings of text.